requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
Brotli==1.1.0
celery==5.3.1
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-debug-toolbar==4.2.0
django-import-export==3.2.0
django-mathfilters==1.0.0
docopt==0.6.2
et-xmlfile==1.1.0
fonttools==4.42.1
html5lib==1.1
kombu==5.3.1
MarkupPy==1.14
num2words==0.5.12
odfpy==1.4.1
openpyxl==3.1.2
Pillow==10.0.1
prompt-toolkit==3.0.39
pycparser==2.21
pydyf==0.7.0
pymemcache==4.0.0
pyphen==0.14.0
python-dateutil==2.8.2
PyYAML==6.0.1
redis==5.0.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.2.1
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.2.3